Created a new class that takes care of the battles
For now it's just the code from the input loop at the input.cpp file,
divided between a couple of functions.

Since the code in the method battle_input() calls the SDL_PollEvent, it
is necessary to check if we are in battle mode, before doing the
"regular" loop on input.cpp

As a side effect of passing the battle code to another file, the option
to see the pokemenu while inside the battle no longer works.

Also, it appears that the player continues walking on the overworld map
while inside a battle, although I'm not sure if this is a new bug or
not.

To see this bug, try entering battle mode while pressing one of the wasd
keys. At the end of the battle, the player will probably be walking
towards a wall. It's also possible to reproduce this bug by walking
horizontally on the player's room and activating the message that
displays "Yeah, the Wii is the best!". After the message disappears the
playing will continue walking in the same direction as before. To stop
this behaviour, simply press again the walking key corresponding to the
same direction that the player is walking.