Animated the life bars
Wrote code that allows the animation of the life bars, unfortunately it may
work too fast to be seen during game play (and adding, for example, a
sleep will also stop the music).

Also the use of the animation appears to cause a bug, in which the game
crashes at the end of a battle with an error.

In order to try and separate the code, I created a method to process
an attack called process_attack.