Corrected game crashing bug at the end of battle
I'm not sure, but I think the bug that caused to game to crash at the
end of a battle has been corrected.

I added a line that makes sure the health of the opponent doesn't drop
into the range of negative numbers, since the remove
little_life_opponent would subtract the value of life_to_remove to the
opponent life bar, even if it would turn into the range of negative
numbers.

The problem was that the bug wasn't solved with this, and it kept
crashing at the end of battle, always something related to the line 70
of input.cpp (newSong("01.ogg")), sometimes crashing somewhere on the
string library, sometimes on some function on the music.cpp file.

Anyway, I commented that line, and the game stopped crashing, although
it wouldn't change back the music at the end of the battle, only
changing when the player changed maps (for example, when the player
walked down the stairs at the beginning of the game).

The weird thing is that I uncommented that line, and the problem didn't
reappear, hence the "I'm not sure I corrected the bug", although I'm
fairly sure this one-line change would need to be done sooner or later
(the life bar "grows" to the left side if there isn't this line).